requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==3.7.1
attrs==25.1.0
beautifulsoup4==4.12.3
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
distro==1.9.0
Faker==36.1.1
fastapi==0.109.2
feedparser==6.0.11
filelock==3.17.0
Flask==3.0.0
fonttools==4.55.8
frozenlist==1.5.0
fsspec==2025.2.0
h11==0.14.0
httpcore==1.0.7
httpx==0.26.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
kiwisolver==1.4.8
llvmlite==0.44.0
Markdown==3.5.2
markdownify==0.14.1
MarkupSafe==3.0.2
matplotlib==3.10.0
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.3
openai==0.28.0
openai-whisper==20240930
packaging==24.2
pandas==2.2.0
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.2.0
propcache==0.2.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
regex==2024.11.6
requests==2.31.0
scikit-learn==1.6.1
scipy==1.12.0
setuptools==75.8.0
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
sympy==1.13.1
threadpoolctl==3.5.0
tiktoken==0.8.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uv==0.5.31
uvicorn==0.27.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.18.3
duckdb==0.9.2
sqlite3==0.0.1
